refactor: convert sign-in page to a server component
