+++ src/app/(auth)/sign-in/page.tsx
@@ -1,24 +1,12 @@
-"use client";
-
-import SignInForm from "@/components/auth/sign-in-form";
-import { useToast } from "@/components/ui/use-toast";
-import useAuthRedirect from "@/hooks/use-auth-redirect";
 import Image from "next/image";
+import SignInForm from "./components/sign-in-form";
 
 export default function SignInPage() {
-  useAuthRedirect(true, "/");
-  const { toast } = useToast();
-
-  const successHandler = () => {
-    toast({ title: "Добро пожаловать!" });
-    window.location.replace("/");
-  };
-
   return (
     <div className="py-10 flex flex-col items-center gap-4">
       <Image src="/images/flat/1.jpg" alt="" width={256} height={256} />
       <h1 className="text-xl font-bold">Вход в личный кабинет</h1>
-      <SignInForm onSuccess={successHandler} />
+      <SignInForm />
     </div>
   );
 }
